Set all money fields as decimal

+++ flightsInfo/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 
 class Package(models.Model):
-    price = models.SmallIntegerField()
+    price = models.DecimalField(max_digits=8, decimal_places=2)
     num_of_flights = models.SmallIntegerField()
     name = models.CharField(max_length=30)
